requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
jinja2
python-multipart